컴포넌트 모듈 분리
