+++ app/src/main/java/com/glowstudio/android/blindsjn/feature/certification/BusinessCertificationScreen.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.glowstudio.android.blindsjn.ui.components.CommonButton
+import com.glowstudio.android.blindsjn.ui.components.common.CommonButton
 
 @Composable
 fun BusinessCertificationScreen(
