+++ app/src/main/java/com/glowstudio/android/blindsjn/feature/foodcoast/RegisterRecipeScreen.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.glowstudio.android.blindsjn.ui.components.CommonButton
+import com.glowstudio.android.blindsjn.ui.components.common.CommonButton
 
 @Composable
 fun RegisterRecipeScreen() {
