+++ app/src/main/java/com/glowstudio/android/blindsjn/ui/components/common/SectionLayout.kt
@@ -1,0 +1,34 @@
+package com.glowstudio.android.blindsjn.ui.components.common
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.KeyboardArrowRight
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun SectionLayout(
+    title: String,
+    onMoreClick: (() -> Unit)? = null,
+    content: @Composable () -> Unit
+) {
+    Column(modifier = Modifier.padding(16.dp)) {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.SpaceBetween,
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(title, fontWeight = FontWeight.Bold)
+            IconButton(onClick = { onMoreClick?.invoke() }) {
+                Icon(Icons.Default.KeyboardArrowRight, contentDescription = "더보기")
+            }
+        }
+
+        Spacer(modifier = Modifier.height(8.dp))
+        content()
+    }
+}
