+++ app/src/main/java/com/glowstudio/android/blindsjn/ui/screens/EditProfileScreen.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.glowstudio.android.blindsjn.ui.components.CommonButton
+import com.glowstudio.android.blindsjn.ui.components.common.CommonButton
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
